chore: manual updates for a clean build. Stubbed migrations.

+++ packages/amplify_datastore/ios/Classes/types/hub/FlutterHubElement.swift
@@ -5,14 +5,14 @@ import Foundation
 import Flutter
 import UIKit
 import Amplify
-import AmplifyPlugins
-import AWSCore
+import AWSDataStorePlugin
 import Combine
 
 public struct FlutterHubElement {
     var model: [String: Any]
     var version: Int?
-    var lastChangedAt: Int?
+    // TODO: Migrate to Async Swift v2 -- Is this breaking?
+    var lastChangedAt: Int64?
     var deleted: Bool
 
     init(
@@ -53,11 +53,11 @@ public struct FlutterHubElement {
             let serializedData = model["serializedData"] as? [String: Any] ?? [:]
             self.deleted = serializedData["_deleted"] as? Bool ?? false
             if let value = serializedData["_lastChangedAt"] as? Double {
-                self.lastChangedAt = Int(value)
+                self.lastChangedAt = Int64(value)
             } else if let value = serializedData["_lastChangedAt"] as? String {
-                self.lastChangedAt = Int(value)
+                self.lastChangedAt = Int64(value)
             } else if let value = serializedData["_lastChangedAt"] as? Int {
-                self.lastChangedAt = value
+                self.lastChangedAt = Int64(value)
             }
         } catch {
             throw FlutterDataStoreError.hubEventCast
